# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Hôtel Artichaut",
        "version": "1.0"
    },
    "paths": {
...
        "/api/header": {
            "get": {
                "tags": [
                    "Header"
                ],
                "summary": "Liste tous les headers avec leurs icônes",
                "operationId": "66b81621bf6ab7a515ebd12b7ff3a2c7",
                "responses": {
                    "200": {
                        "description": "Liste des headers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Header"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Header"
                ],
                "summary": "Créer un nouveau header",
                "operationId": "b153c1cd451a7cba14921e776abf74d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Header"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Header créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Header"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/header/{id}": {
            "get": {
                "tags": [
                    "Header"
                ],
                "summary": "Afficher un header par ID",
                "operationId": "8dc075e61109c418fecc9cbeab7836db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Header trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Header"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Header"
                ],
                "summary": "Mettre à jour un header",
                "operationId": "d6889bfd03a8086f94aa4cb6f4a78eba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Header"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Header mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Header"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Header"
                ],
                "summary": "Supprimer un header",
                "operationId": "ba0a375581660600afbd41976b3a10c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Header supprimé, retourne tous les headers restants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Header"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/icons": {
            "get": {
                "tags": [
                    "Icons"
                ],
                "summary": "Liste toutes les icônes",
                "operationId": "e570f65e486c2352fddd792592da3df1",
                "responses": {
                    "200": {
                        "description": "Liste des icônes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Icons"
                ],
                "summary": "Crée une nouvelle icône",
                "operationId": "a5468a4e83d3ad2a82fa438cfb64b933",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Icône créée"
                    }
                }
            }
        },
        "/api/icons/{id}": {
            "get": {
                "tags": [
                    "Icons"
                ],
                "summary": "Affiche une icône par ID",
                "operationId": "605d5119c4b301bfeb7d1db51a8b6ebe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Icône trouvée"
                    },
                    "404": {
                        "description": "Icône non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Icons"
                ],
                "summary": "Met à jour une icône",
                "operationId": "d71f5994baf591193a26c26d5e537191",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Icône mise à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Icons"
                ],
                "summary": "Supprime une icône",
                "operationId": "c2e9cc482599e9b7b31883864574724a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Icône supprimée, retourne la liste restante"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Header": {
                "title": "Header",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "logo": {
                        "type": "string",
                        "example": "logo.png"
                    },
                    "icone": {
                        "type": "string",
                        "example": "icon.png"
                    },
                    "background_color_1": {
                        "type": "string",
                        "example": "#ffffff"
                    },
                    "background_opacity_1": {
                        "type": "integer",
                        "example": 80
                    },
                    "fondus_color_2": {
                        "type": "string",
                        "example": "#000000"
                    },
                    "fondus_opacity_2": {
                        "type": "integer",
                        "example": 50
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-24T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-24T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}